+++ rebound/particle.py
@@ -49,10 +49,10 @@ class Particle(Structure):
         """ 
         Returns a string with the position and velocity of the particle.
         """
-        return '<{0}.{1} object at {2}, m={3} x={4} y={5} z={6} vx={7} vy={8} vz={8}>'.format(self.__module__, type(self).__name__, hex(id(self)), self.m, self.x, self.y, self.z, self.vx, self.vy, self.vz)
+        return '<{0}.{1} object at {2}, m={3} x={4} y={5} z={6} vx={7} vy={8} vz={9}>'.format(self.__module__, type(self).__name__, hex(id(self)), self.m, self.x, self.y, self.z, self.vx, self.vy, self.vz)
    
 
-    def __init__(self, simulation=None, particle=None, m=None, x=None, y=None, z=None, vx=None, vy=None, vz=None, primary=None, a=None, P=None, e=None, inc=None, Omega=None, omega=None, pomega=None, f=None, M=None, l=None, theta=None, T=None, r=None, date=None, variation=None, variation2=None, h=None, k=None, ix=None, iy=None, hash=0, jacobi_masses=False):
+    def __init__(self, simulation=None, particle=None, m=None, x=None, y=None, z=None, vx=None, vy=None, vz=None, primary=None, a=None, P=None, e=None, inc=None, Omega=None, omega=None, pomega=None, f=None, M=None, E=None, l=None, theta=None, T=None, r=None, date=None, variation=None, variation2=None, h=None, k=None, ix=None, iy=None, hash=0, jacobi_masses=False):
         """
         Initializes a Particle structure. Rather than explicitly creating 
         a Particle structure, users may use the ``add()`` member function 
@@ -114,6 +114,8 @@ class Particle(Structure):
             True anomaly                (Default: 0)
         M           : float       
             Mean anomaly                (Default: 0)
+        E           : float       
+            Eccentric anomaly           (Default: 0)
         l           : float       
             Mean longitude              (Default: 0)
         theta       : float       
@@ -164,6 +166,8 @@ class Particle(Structure):
             f = random.vonmisesvariate(0.,0.) 
         if M == "uniform":
             M = random.vonmisesvariate(0.,0.) 
+        if E == "uniform":
+            E = random.vonmisesvariate(0.,0.) 
         if l == "uniform":
             l = random.vonmisesvariate(0.,0.) 
         if theta == "uniform":
@@ -234,7 +238,7 @@ class Particle(Structure):
             memmove(byref(self), byref(particle), sizeof(self))
             return
         cart = [x,y,z,vx,vy,vz]
-        orbi = [primary,a,P,e,inc,Omega,omega,pomega,f,M,l,theta,T]
+        orbi = [primary,a,P,e,inc,Omega,omega,pomega,f,M,E,l,theta,T]
         pal  = [h,k,ix,iy]
        
         self.ax = 0.
@@ -252,8 +256,8 @@ class Particle(Structure):
         self.c = None
         self.ap = None
         
-        if notNone([e,inc,omega,pomega,Omega,M,f,theta,T]) and notNone(pal):
-            raise ValueError("You cannot mix Pal coordinates (h,k,ix,iy) with the following orbital elements: e,inc,Omega,omega,pomega,f,M,theta,T. If a longitude/anomaly is needed in Pal coordinates, use l.")
+        if notNone([e,inc,omega,pomega,Omega,M,f,E,theta,T]) and notNone(pal):
+            raise ValueError("You cannot mix Pal coordinates (h,k,ix,iy) with the following orbital elements: e,inc,Omega,omega,pomega,f,M,E,theta,T. If a longitude/anomaly is needed in Pal coordinates, use l.")
         if notNone(cart) and notNone(orbi):
                 raise ValueError("You cannot pass cartesian coordinates and orbital elements (and/or primary) at the same time.")
         if notNone(orbi):
@@ -313,33 +317,40 @@ class Particle(Structure):
                         else:
                             omega = Omega - pomega  # for retrograde orbits, pomega = Omega - omega
 
-                longitudes = [f,M,l,theta,T]      # can specify longitude through any of these four.  Need f for C function.
+                longitudes = [f,M,E,l,theta,T]      # can specify longitude through any of these.  Need f for C function.
                 numNones = longitudes.count(None)
 
-                if numNones < 4:
-                    raise ValueError("Can only pass one longitude/anomaly in the set [f, M, l, theta, T]")
-                if numNones == 5:                           # none of them passed.  Default to 0.
+                if numNones < 5:
+                    raise ValueError("Can only pass one longitude/anomaly in the set [f, M, E, l, theta, T]")
+                if numNones == 6:                           # none of them passed.  Default to 0.
                     f = 0.
-                if numNones == 4:                           # Only one was passed.
-                    if f is None:                           # Only have to work if f wasn't passed.
-                        if theta is not None:               # theta is next easiest
-                            if math.cos(inc) > 0:           # for prograde orbits, theta = Omega + omega + f
-                                f = theta - Omega - omega
-                            else:
-                                f = Omega - omega - theta   # for retrograde, theta = Omega - omega - f
-                        else:                               # Either M, l, or T was passed.  Will need to find M first (if not passed) to find f
-                            if l is not None:
-                                if math.cos(inc) > 0:       # for prograde orbits, l = Omega + omega + M
-                                    M = l - Omega - omega
-                                else:
-                                    M = Omega - omega - l   # for retrograde, l = Omega - omega - M
-                            else:
-                                if T is not None:           # works for both elliptical and hyperbolic orbits
-                                                            # TODO: has accuracy problems for M=n*(t-T) << 1
-                                    n = (simulation.G*(primary.m+self.m)/abs(a**3))**0.5
-                                    M = n*(simulation.t - T)
-                            clibrebound.reb_tools_M_to_f.restype = c_double
-                            f = clibrebound.reb_tools_M_to_f(c_double(e), c_double(M))
+                if numNones == 5:                           # Only one was passed. Calculate f.
+                    if f is not None:
+                        pass                            # Nothing to be done
+                    elif theta is not None:               # theta is next easiest
+                        if math.cos(inc) > 0:           # for prograde orbits, theta = Omega + omega + f
+                            f = theta - Omega - omega
+                        else:
+                            f = Omega - omega - theta   # for retrograde, theta = Omega - omega - f
+                    elif l is not None:                 
+                        if math.cos(inc) > 0:           # for prograde orbits, l = Omega + omega + M
+                            M = l - Omega - omega
+                        else:
+                            M = Omega - omega - l       # for retrograde, l = Omega - omega - M
+                        clibrebound.reb_tools_M_to_f.restype = c_double
+                        f = clibrebound.reb_tools_M_to_f(c_double(e), c_double(M))
+                    elif T is not None:                 # works for both elliptical and hyperbolic orbits
+                                                        # TODO: has accuracy problems for M=n*(t-T) << 1
+                        n = (simulation.G*(primary.m+self.m)/abs(a**3))**0.5
+                        M = n*(simulation.t - T)
+                        clibrebound.reb_tools_M_to_f.restype = c_double
+                        f = clibrebound.reb_tools_M_to_f(c_double(e), c_double(M))
+                    elif M is not None:
+                        clibrebound.reb_tools_M_to_f.restype = c_double
+                        f = clibrebound.reb_tools_M_to_f(c_double(e), c_double(M))
+                    elif E is not None:
+                        clibrebound.reb_tools_E_to_f.restype = c_double
+                        f = clibrebound.reb_tools_E_to_f(c_double(e), c_double(E))
 
                 err = c_int()
                 clibrebound.reb_tools_orbit_to_particle_err.restype = Particle
@@ -448,7 +459,7 @@ class Particle(Structure):
 
         return o
     
-    def sample_orbit(self, Npts=100, primary=None, trailing=True, timespan=None, useTrueAnomaly=None, duplicateEndpoint=True):
+    def sample_orbit(self, Npts=100, primary=None, samplingAngle=None, duplicateEndpoint=None):
         """
         Returns a nested list of xyz positions along the osculating orbit of the particle. 
         If primary is not passed, returns xyz positions along the Jacobi osculating orbit
@@ -460,52 +471,101 @@ class Particle(Structure):
             Number of points along the orbit to return  (default: 100)
         primary : rebound.Particle, optional
             Primary to use for the osculating orbit (default: Jacobi center of mass)
-        trailing: bool, optional
-            Whether to return points stepping backwards in time (True) or forwards (False). (default: True)
-        timespan: float, optional    
-            Return points (for the osculating orbit) from the current position to timespan (forwards or backwards in time depending on trailing keyword). 
-            Defaults to the orbital period for bound orbits, and to the rough time it takes the orbit to move by the current distance from the primary for a hyperbolic orbit. Implementation currently only supports this option if useTrueAnomaly=False.
-        useTrueAnomaly: bool, optional
-            Will sample equally spaced points in true anomaly if True, otherwise in mean anomaly.
-            Latter might be better for hyperbolic orbits, where true anomaly can stay near the limiting value for a long time, and then switch abruptly at pericenter. (Default: True for bound orbits, False for unbound orbits)
+        samplingAngle: str, optional
+            This determines which angle is sampled linearly. Can be:
+              - "M" (mean anomaly)
+              - "E" (eccentric anomaly) 
+              - "f" (true anomaly)
         duplicateEndpoint: bool, optional
-            If true (default), then the first and last point will be identical for closed orbits. This is useful for some plotting tools.
+            If true, then the first and last point will be identical for closed orbits. This is useful for some plotting tools. The default is true for eccentric orbits. The argument has no effect for hyperbolic orbits (because the endpoints are not identical).
         """
-        pts = []
         if primary is None:
             primary = self.jacobi_com
         o = self.calculate_orbit(primary=primary)
 
-        if timespan is None:
-            if o.a < 0.: # hyperbolic orbit
-                timespan = 2*math.pi*o.d/o.v # rough time to cross display box
-            else:
-                timespan = o.P
+        phases_f = []
+        clibrebound.reb_tools_E_to_f.restype = c_double
+        clibrebound.reb_tools_M_to_f.restype = c_double
+        clibrebound.reb_tools_mod2pi.restype = c_double
+        if samplingAngle is not None:
+            if any(c not in "EMf" for c in samplingAngle):
+                raise ValueError("Unknown character in samplingAngle.")
         
-        lim_phase = abs(o.n)*timespan # n is negative for hyperbolic orbits
+        if o.a < 0.: # hyperbolic orbit
+            if samplingAngle is None:
+                samplingAngle = "Mf"
+            Nptsangle = {}
+            for angle in samplingAngle[1:]:
+                Nptsangle[angle] = (Npts-1)//len(samplingAngle) # one point is reserved for actual position
+            Nptsangle[samplingAngle[0]] = Npts-1-sum(Nptsangle.values())
+            if "M" in samplingAngle:
+                phi = math.acos(-1./o.e)*0.999
+                Npts = Nptsangle["M"]
+                dphi = 2*phi/(Npts-1)
+                for i in range(Npts):
+                    f = clibrebound.reb_tools_M_to_f(c_double(o.e), c_double(phi))
+                    phases_f.append(f)
+                    phi -= dphi
+            if "E" in samplingAngle:
+                phi = math.acos(-1./o.e)*0.999
+                Npts = Nptsangle["E"]
+                dphi = 2*phi/(Npts-1)
+                for i in range(Npts):
+                    f = clibrebound.reb_tools_E_to_f(c_double(o.e), c_double(phi))
+                    phases_f.append(f)
+                    phi -= dphi
+            if "f" in samplingAngle:
+                phi = math.acos(-1./o.e)*0.999
+                Npts = Nptsangle["f"]
+                dphi = 2*phi/(Npts-1)
+                for i in range(Npts):
+                    f = clibrebound.reb_tools_mod2pi(c_double(phi))
+                    phases_f.append(f)
+                    phi -= dphi
+        else:       # circular orbit
+            if samplingAngle is None:
+                samplingAngle = "Ef"
+            if duplicateEndpoint is None:
+                duplicateEndpoint = True
+            Nptsangle = {}
+            for angle in samplingAngle[1:]:
+                Nptsangle[angle] = (Npts-1)//len(samplingAngle) # one point is reserved for actual position
+            Nptsangle[samplingAngle[0]] = Npts-1-sum(Nptsangle.values())
+            if "M" in samplingAngle:
+                Npts = Nptsangle["M"]
+                dphi = 2.*math.pi/(Npts-1 if duplicateEndpoint else Npts)  # one point is reserved for the end point
+                for i in range(Npts):
+                    f = clibrebound.reb_tools_M_to_f(c_double(o.e), c_double(i*dphi))
+                    phases_f.append(f)
+            if "E" in samplingAngle:
+                Npts = Nptsangle["E"]
+                dphi = 2.*math.pi/(Npts-1 if duplicateEndpoint else Npts)  # one point is reserved for the end point
+                for i in range(Npts):
+                    f = clibrebound.reb_tools_E_to_f(c_double(o.e), c_double(i*dphi))
+                    phases_f.append(f)
+            if "f" in samplingAngle:
+                Npts = Nptsangle["f"]
+                dphi = 2.*math.pi/(Npts-1 if duplicateEndpoint else Npts)  # one point is reserved for the end point
+                for i in range(Npts):
+                    f = i*dphi
+                    f = clibrebound.reb_tools_mod2pi(c_double(f))
+                    phases_f.append(f)
 
-        if trailing is True:
-            lim_phase *= -1 # sample phase backwards from current value
-        _Npts = Npts
-        if duplicateEndpoint:
-            _Npts -= 1
-        phase = [lim_phase*i/_Npts  for i in range(Npts)]
-
-        if useTrueAnomaly is None:
-            if o.a <0.:
-                useTrueAnomaly = False
+        # add actual position
+        f = clibrebound.reb_tools_mod2pi(c_double(o.f))
+        phases_f.append(f)
+        phases_f.sort()
+      
+        pts_pre = []
+        pts_post = []
+        for f in phases_f:
+            newp = Particle(a=o.a, f=f, inc=o.inc, omega=o.omega, Omega=o.Omega, e=o.e, m=self.m, primary=primary, simulation=self._sim.contents)
+            if f<=o.f:
+                pts_pre.append(newp.xyz)
             else:
-                useTrueAnomaly = True
-
-        for i,ph in enumerate(phase):
-            if useTrueAnomaly is True:
-                newp = Particle(a=o.a, f=o.f+ph, inc=o.inc, omega=o.omega, Omega=o.Omega, e=o.e, m=self.m, primary=primary, simulation=self._sim.contents)
-            else: 
-                newp = Particle(a=o.a, M=o.M+ph, inc=o.inc, omega=o.omega, Omega=o.Omega, e=o.e, m=self.m, primary=primary, simulation=self._sim.contents)
-
-            pts.append(newp.xyz)
+                pts_post.append(newp.xyz)
         
-        return pts
+        return pts_post + pts_pre
 
     # Simple operators for particles.
     
@@ -633,45 +693,153 @@ class Particle(Structure):
     @property
     def P(self):
         return self.calculate_orbit().P
+    @P.setter
+    def P(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, P=value, e=o.e, inc=o.inc, omega=o.omega, Omega=o.Omega, f=o.f) 
+        self._cpcoords(newP)
     @property
     def n(self):
         return self.calculate_orbit().n 
     @property
     def a(self):
         return self.calculate_orbit().a 
+    @a.setter
+    def a(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=value, e=o.e, inc=o.inc, omega=o.omega, Omega=o.Omega, f=o.f) 
+        self._cpcoords(newP)
     @property
     def rhill(self):
         return self.calculate_orbit().rhill
     @property
     def e(self):
         return self.calculate_orbit().e 
+    @e.setter
+    def e(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=o.a, e=value, inc=o.inc, omega=o.omega, Omega=o.Omega, f=o.f) 
+        self._cpcoords(newP)
     @property
     def inc(self):
         return self.calculate_orbit().inc 
+    @inc.setter
+    def inc(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=o.a, e=o.e, inc=value, omega=o.omega, Omega=o.Omega, f=o.f) 
+        self._cpcoords(newP)
     @property
     def Omega(self):
         return self.calculate_orbit().Omega 
+    @Omega.setter
+    def Omega(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=o.a, e=o.e, inc=o.inc, omega=o.omega, Omega=value, f=o.f) 
+        self._cpcoords(newP)
     @property
     def omega(self):
         return self.calculate_orbit().omega 
+    @omega.setter
+    def omega(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=o.a, e=o.e, inc=o.inc, omega=value, Omega=o.Omega, f=o.f) 
+        self._cpcoords(newP)
     @property
     def pomega(self):
         return self.calculate_orbit().pomega 
+    @pomega.setter
+    def pomega(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=o.a, e=o.e, inc=o.inc, pomega=value, Omega=o.Omega, f=o.f) 
+        self._cpcoords(newP)
     @property
     def f(self):
         return self.calculate_orbit().f 
+    @f.setter
+    def f(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=o.a, e=o.e, inc=o.inc, omega=o.omega, Omega=o.Omega, f=value) 
+        self._cpcoords(newP)
     @property
     def M(self):
         return self.calculate_orbit().M 
+    @M.setter
+    def M(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=o.a, e=o.e, inc=o.inc, omega=o.omega, Omega=o.Omega, M=value) 
+        self._cpcoords(newP)
     @property
     def l(self):
         return self.calculate_orbit().l 
+    @l.setter
+    def l(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=o.a, e=o.e, inc=o.inc, omega=o.omega, Omega=o.Omega, l=value) 
+        self._cpcoords(newP)
     @property
     def theta(self):
         return self.calculate_orbit().theta 
+    @theta.setter
+    def theta(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=o.a, e=o.e, inc=o.inc, omega=o.omega, Omega=o.Omega, theta=value) 
+        self._cpcoords(newP)
     @property
     def T(self):
         return self.calculate_orbit().T
+    @T.setter
+    def T(self,value):
+        o = self.calculate_orbit()
+        clibrebound.reb_get_jacobi_com.restype = Particle
+        primary = clibrebound.reb_get_jacobi_com(byref(self))
+        if self._sim is None:
+            raise RuntimeError("Cannot modify particle which is not a member of a simulation.")
+        newP = Particle(simulation=self._sim.contents, primary=primary, m=self.m, a=o.a, e=o.e, inc=o.inc, omega=o.omega, Omega=o.Omega, T=value) 
+        self._cpcoords(newP)
     @property
     def orbit(self):
         return self.calculate_orbit()
@@ -703,3 +871,10 @@ class Particle(Structure):
             self._hash = value
         else:
             raise AttributeError("Hash must be set to an integer, a ctypes.c_uint32 or a string. See UniquelyIdentifyingParticlesWithHashes.ipynb ipython_example.")
+
+    def _cpcoords(self, p):
+        """
+        Copy coordinates (and only coordinates) from particle p to self
+        """
+        self.xyz = p.xyz
+        self.vxyz = p.vxyz