+++ src/input.c
@@ -451,6 +451,9 @@ struct reb_simulation* reb_input_process_warnings(struct reb_simulation* r, enum
         if (r) free(r);
         return NULL;
     }
+    if (warnings & REB_INPUT_BINARY_WARNING_CORRUPTFILE){
+        reb_warning(r,"The binary file seems to be corrupted. An attempt has been made to recover parts of it. However, it might not be possible to append snapshots to the current file.");
+    }
     return r;
 }