+++ src/rebound.h
@@ -271,6 +271,7 @@ struct reb_simulation_integrator_mercurius {
     unsigned int mode;              ///< Internal. 0 if WH is operating, 1 if IAS15 is operating.
     unsigned int encounterN;        ///< Number of particles currently having an encounter
     unsigned int encounterNactive;  ///< Number of particles currently having an encounter
+    unsigned int tponly_encounter;  ///< Flag to determine if any of the encounters are between two massive bodies (0) or only involve test particles (1). Internal use only.
     unsigned int allocatedN;        ///< Current size of allocated internal arrays
     unsigned int allocatedN_additionalforces;        ///< Current size of allocated internal particles_backup_additionalforces array
     unsigned int dcrit_allocatedN;  ///< Current size of dcrit arrays
@@ -1550,6 +1551,12 @@ void reb_output_velocity_dispersion(struct reb_simulation* r, char* filename);
  * List of the built-in setup helper functions
  * @{
  */
+
+/**
+ * @brief Returns f modulo 2 pi. Return value is in the range [0, 2pi)
+ */
+double reb_tools_mod2pi(double f);
+
 /**
  * @brief returns the true anomaly for a given eccentricity and mean anomaly
  * @param e Eccentricity
@@ -1669,6 +1676,7 @@ enum reb_input_binary_messages {
     REB_INPUT_BINARY_ERROR_SEEK = 64,
     REB_INPUT_BINARY_WARNING_FIELD_UNKOWN = 128,
     REB_INPUT_BINARY_ERROR_INTEGRATOR = 256,
+    REB_INPUT_BINARY_WARNING_CORRUPTFILE = 512,
 };
 
 /**